Split data.json into config.json and data.csv
