+++ config.json
@@ -1,0 +1,4 @@
+{
+  "currentLevel": 0,
+  "semesterAccount": "nus-cs2103-AY1718S1"
+}
