--- src/data.json
@@ -1,77 +1,0 @@
-{
-  "currentLevel": 0,
-  "semesterAccount": "nus-cs2103-AY1718S1",
-  "classes": {
-    "W09": {
-      "1": "jeffryhartanto",
-      "2": "jeffryhartanto",
-      "3": "hwkchia",
-      "4": "hwkchia"
-    },
-    "W10": {
-      "1": "sooyj",
-      "2": "sooyj",
-      "3": "hwkchia",
-      "4": "hwkchia"
-    },
-    "W13": {
-      "1": "sooyj",
-      "2": "sooyj",
-      "3": "K1ang",
-      "4": "K1ang"
-    },
-    "W14": {
-      "1": "mauris",
-      "2": "mauris",
-      "3": "K1ang",
-      "4": "K1ang"
-    },
-    "W15": {
-      "1": "mauris",
-      "2": "mauris",
-      "3": "cmkumar87",
-      "4": "cmkumar87"
-    },
-    "T09": {
-      "1": "nirandiw",
-      "2": "nirandiw",
-      "3": "okkhoy",
-      "4": ["okkhoy", "ZawadulFarhan"]
-    },
-    "T11": {
-      "1": "nirandiw",
-      "2": "nirandiw",
-      "3": "ZawadulFarhan",
-      "4": "ZawadulFarhan"
-    },
-    "T15": {
-      "1": "okkhoy",
-      "2": ["okkhoy", "xpdavid"],
-      "3": "joelt-92",
-      "4": "joelt-92"
-    },
-    "T16": {
-      "1": "okkhoy",
-      "2": ["okkhoy", "brandonyeoxg"],
-      "3": "joelt-92",
-      "4": "joelt-92"
-    },
-    "F11": {
-      "1": "cmkumar87",
-      "2": "cmkumar87",
-      "3": "xpdavid",
-      "4": "xpdavid"
-    },
-    "F12": {
-      "1": "brandonyeoxg",
-      "2": "brandonyeoxg",
-      "3": "okkhoy",
-      "4": "okkhoy"
-    },
-    "F14": {
-      "1": "chanjunweimy",
-      "2": "chanjunweimy",
-      "3": "nirandiw"
-    }
-  }
-}
